Split server so we can run  integration tests.
