+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node server.js",
+    "start": "node start.js",
     "lint": "eslint . --ext .js",
     "test": "jest"
   },
