+++ server.js
@@ -1,10 +1,8 @@
 const express = require('express');
 const app = express();
 const morgan = require('morgan');
-const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
 const cors = require('cors');
-const { mongodbUri } = require('./util');
 const { bugsnagMiddleware } = require('./bugsnag');
 
 // enable cors in preflight
@@ -19,11 +17,6 @@ app.use('/public', express.static(__dirname + '/public'));
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
 app.use(morgan('short'));
-// app.use(function(req, res, next) {
-//     res.header("Access-Control-Allow-Origin", "*");
-//     res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
-//     next();
-// });
 app.use(cors());
 
 // Register routes
@@ -53,16 +46,4 @@ app.use(function(req, res, _next) {
 
 app.use(bugsnagMiddleware.errorHandler);
 
-// Connect to database and start the serverfuser
-mongoose.connect(mongodbUri, (err, _database) => {
-  if (err) {
-    console.log(err);
-    process.exit(1);
-  }
-  console.log('Database connection ready');
-
-  const server = app.listen(process.env.PORT || 3000, () => {
-    const port = server.address().port;
-    console.log(`Listening on port ${port}!`);
-  });
-});
+module.exports = app;
